+++ src/particles/image/component.tsx
@@ -88,7 +88,8 @@ export const Image = (props: IImageProps): React.ReactElement => {
   const fitType = props.fitType || 'scale';
   const width = props.width ? props.width : props.isFullWidth ? '100%' : 'auto';
   const height = props.height ? props.height : props.isFullHeight ? '100%' : 'auto';
-  const isSourceResponsive = props.source.includes('d35ci2i0uce4j6.cloudfront.net') || props.source.includes('pablo-images.kibalabs.com');
+  const source = props.source.startsWith('ipfs://') ? props.source.replace('ipfs://', 'https://pablo-images.kibalabs.com/v1/ipfs/') : props.source;
+  const isSourceResponsive = source.includes('d35ci2i0uce4j6.cloudfront.net') || source.includes('pablo-images.kibalabs.com');
 
   return (
     <React.Fragment>
@@ -101,12 +102,12 @@ export const Image = (props: IImageProps): React.ReactElement => {
         $height={height}
         $maxWidth={props.maxWidth || 'none'}
         $maxHeight={props.maxHeight || 'none'}
-        src={props.isLazyLoadable ? undefined : props.source}
-        data-src={props.isLazyLoadable ? props.source : undefined}
+        src={props.isLazyLoadable ? undefined : source}
+        data-src={props.isLazyLoadable ? source : undefined}
         sizes={isSourceResponsive && !props.isLazyLoadable ? 'auto' : undefined}
         data-sizes={isSourceResponsive && props.isLazyLoadable ? 'auto' : undefined}
-        srcSet={isSourceResponsive && !props.isLazyLoadable ? getResponsiveImageString(props.source) : undefined}
-        data-srcset={isSourceResponsive && props.isLazyLoadable ? getResponsiveImageString(props.source) : undefined}
+        srcSet={isSourceResponsive && !props.isLazyLoadable ? getResponsiveImageString(source) : undefined}
+        data-srcset={isSourceResponsive && props.isLazyLoadable ? getResponsiveImageString(source) : undefined}
         alt={props.alternativeText}
       />
       {props.isLazyLoadable && (